requirements.txt
langchain-openai>=0.1.0
browser-use>=0.1.0
python-dotenv>=1.0.0 